requirements.txt
django
psycopg2-binary
Pillow>=8.0.0
channels
gunicorn>=20.1.0
djangorestframework


